requirements.txt
beautifulsoup4==4.13.4
cobble==0.1.4
lxml==6.0.0
mammoth==1.9.1
markdownify==1.1.0
pillow==11.3.0
python-docx==1.2.0
six==1.17.0
soupsieve==2.7
typing_extensions==4.14.1
